move AdTokenToWallet component into BridgeForm
